bus-launch: add dbus-broker support
If dbus-broker is enabled at compile-time and found at run-time prefer that
over dbus-daemon.

Unlike with dbus-daemon, this means that at-spi-bus-launch must create the
listening socket and pass it in, rather than having the bus do that and send
back the address. For now we follow what dbus-daemon does, and create a socket
in the abstract namespace, though it might be more suitable to create a socket
in $XDG_RUNTIME_DIR.

The only difference users should observe is that daemons are no longer spawned
by the bus implementation, but spawned and managed by the systemd user instance,
though this should not lead to a difference in behavior. In particular this
applies to `org.a11y.atspi.Registry`.

For non-linux and non-systemd systems, dbus-daemon should continue to be used.

Signed-off-by: Tom Gundersen <teg@jklm.no>